[fix](cast) fix implicit casts in disk_info

+++ be/src/util/disk_info.cpp
@@ -40,6 +40,8 @@
 
 namespace doris {
 
+#include "common/compile_check_begin.h"
+
 bool DiskInfo::_s_initialized;
 std::vector<DiskInfo::Disk> DiskInfo::_s_disks;
 std::map<dev_t, int> DiskInfo::_s_device_id_to_disk_id;
@@ -87,7 +89,7 @@ void DiskInfo::get_device_names() {
 
         if (it == _s_disk_name_to_disk_id.end()) {
             // First time seeing this disk
-            disk_id = _s_disks.size();
+            disk_id = static_cast<int>(_s_disks.size());
             _s_disks.push_back(Disk(name, disk_id));
             _s_disk_name_to_disk_id[name] = disk_id;
         } else {
@@ -231,5 +233,7 @@ Status DiskInfo::get_disk_devices(const std::vector<std::string>& paths,
     fclose(fp);
     return status;
 }
+
+#include "common/compile_check_end.h"
 
 } // namespace doris
